One more cast issue id'd by windows

+++ wolfcrypt/test/test.c
@@ -25416,7 +25416,7 @@ static int ecc_encrypt_e2e_test(WC_RNG* rng, ecc_key* userA, ecc_key* userB,
 
     /* msg2 (response) from B to A */
     for (i = 0; i < (int)sizeof(msg2); i++)
-        msg2[i] = i + sizeof(msg2);
+        msg2[i] = (byte) (i + sizeof(msg2));
 
     /* get encrypted msg (response) to send to B */
     ret = wc_ecc_encrypt(userB, userA, msg2, sizeof(msg2), out2,
